% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{add_external_resources}
\alias{add_external_resources}
\title{Attach external resources}
\usage{
add_external_resources()
}
\description{
Attaches external resources to the head tag, used
to construct user interface
}
