% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{add_orgnr}
\alias{add_orgnr}
\title{adds HF or RHF organization nr}
\usage{
add_orgnr(data_list, by = "RHF")
}
\arguments{
\item{data_list}{a list of dataframe}

\item{by}{can be RHF or HF}
}
\value{
the input data with organization nr appended
}
\description{
adds HF or RHF organization nr
}
