% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aggr_data}
\alias{aggr_data}
\title{Data used by the app if database is not used}
\format{
An object of class \code{list} of length 5.
}
\usage{
aggr_data
}
\description{
Data used by the app if database is not used
}
\keyword{datasets}
