% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{aggregate_data}
\alias{aggregate_data}
\title{aggregate data}
\usage{
aggregate_data(data_list)
}
\arguments{
\item{data_list}{a list of dataframes}
}
\value{
aggregated data
}
\description{
aggregate data
}
