% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{app_data}
\alias{app_data}
\title{makes the data ready to be used by the app}
\usage{
app_data(config)
}
\arguments{
\item{config}{configuration file}
}
\value{
input data to the html table in the app
}
\description{
makes the data ready to be used by the app
}
