% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{compute_indicator}
\alias{compute_indicator}
\title{calculates ki (portion in \% )}
\usage{
compute_indicator(agg_data)
}
\arguments{
\item{agg_data}{aggregated data}
}
\value{
aggrgated data with indicators
}
\description{
calculates ki (portion in \% )
}
