% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{compute_indicator_mean}
\alias{compute_indicator_mean}
\title{calculates  mean value of the grouped variabels}
\usage{
compute_indicator_mean(grouped_data)
}
\arguments{
\item{grouped_data}{aggregated data}
}
\value{
aggrgated data with indicators
}
\description{
calculates  mean value of the grouped variabels
}
