% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{compute_national_indicator}
\alias{compute_national_indicator}
\title{calculates the values of national KI}
\usage{
compute_national_indicator(kidata)
}
\arguments{
\item{kidata}{a dataframe}
}
\value{
dataframe with the the values of national ki
}
\description{
calculates the values of national KI
}
