% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fagomr}
\alias{fagomr}
\title{Medical fields and registries that belongs to these medical fields}
\format{
An object of class \code{list} of length 18.
}
\usage{
fagomr
}
\description{
Medical fields and registries that belongs to these medical fields
}
\keyword{datasets}
