% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{filter_data}
\alias{filter_data}
\title{filter data}
\usage{
filter_data(data_list, filter_settings = NULL)
}
\arguments{
\item{data_list}{a list of dataframes}

\item{filter_settings}{list of values that will be used to subset the
dataframe. The list elements should have the same name as the columns
in the dataframe.}
}
\value{
list of filtered dataframes
}
\description{
filter data
}
