% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_config}
\alias{get_config}
\title{Retrieve Configuration}
\usage{
get_config(dir = ".")
}
\arguments{
\item{dir}{Folder location of _qmongr.yml file}
}
\description{
Retrieve configuration file, checks it first
}
