% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{get_indicator_level}
\alias{get_indicator_level}
\title{Adds a column with inidcator levels and
and another with desired levels to the grouped data}
\usage{
get_indicator_level(grouped_data, description)
}
\arguments{
\item{grouped_data}{a dataframe}

\item{description}{a}
}
\value{
the input data with ad
ditional indicator level
and desired level columns
}
\description{
Adds a column with inidcator levels and
and another with desired levels to the grouped data
}
