% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{group_data}
\alias{group_data}
\title{group data}
\usage{
group_data(data_list, by)
}
\arguments{
\item{data_list}{a list of dataframe}

\item{by}{can be RHF, HF, hospital og an empty string}
}
\value{
aggregated data
}
\description{
group data
}
