% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{has_config}
\alias{has_config}
\title{Has Config}
\usage{
has_config(config_file)
}
\arguments{
\item{config_file}{Config file to ensure is present}
}
\description{
Ensure config file is present.
}
\keyword{internal}
