% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{icon_type}
\alias{icon_type}
\title{adds an icon to the table cells with qi}
\usage{
icon_type(level)
}
\arguments{
\item{level}{the indicator level}
}
\description{
adds an icon to the table cells with qi
}
