% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{indicator_rows}
\alias{indicator_rows}
\title{adds table rows with the values of quality indicators}
\usage{
indicator_rows(indicator_name, indicator_description, config, datatable)
}
\arguments{
\item{indicator_name}{names of the input qi}

\item{indicator_description}{indicator description}

\item{config}{the configuration setup}
}
\description{
adds table rows with the values of quality indicators
}
