% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national_table.R
\name{indicator_value_td}
\alias{indicator_value_td}
\title{adds td tag around indicator values}
\usage{
indicator_value_td(td_data, col_name_indicator_id, class_name = "National")
}
\arguments{
\item{td_data}{a row of indicator data from the dataframe that
corresponds to the selected unit}

\item{col_name_indicator_id}{column name of the indicator column
from config}

\item{class_name}{class name of the td. can be "selected_unit" or "National"}
}
\value{
td with two divs. The first includes the indicator value
and corresponding icon and the second the number of individuals
that make up the numbers.
}
\description{
adds a td tag around the indcator values, level and
N of a selected unit and national values
}
