% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{ki_table}
\alias{ki_table}
\title{html table}
\usage{
ki_table(table_data, enhet)
}
\arguments{
\item{table_data}{a dataframe to construct a html table from}

\item{enhet}{the chosen unit}
}
\description{
html table
}
