% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{load_data}
\alias{load_data}
\title{load data}
\usage{
load_data(data_type = "begge")
}
\arguments{
\item{data_type}{the dataframe to be loaded.
It can be "beskrivelse", "indikator" or "begge".}
}
\value{
a list with dataframes
}
\description{
load data
}
\examples{
indkator_beskrivelse <- load_data(data_type = "beskrivelse")
indkator_data <- load_data(data_type = "indikator")
indikator_med_beskrivelse <- load_data()

}
