% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_field_list.R
\name{med_field_list}
\alias{med_field_list}
\alias{qi_overview}
\title{Clickable links list}
\usage{
med_field_list(id, category_name, nr_of_reg, all_id)
}
\arguments{
\item{id}{unique id for each link}

\item{category_name}{the names of a category that will
be displayed}

\item{nr_of_reg}{the nr of registers included in a category}

\item{all_id}{the inputId for the Alle actionLink}
}
\value{
html tags
}
\description{
Makes html list of clickable links, used to make the sidebar
quality overview user interface
}
\seealso{
\code{\link{sidebar_qo_ui}}
}
