% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_kvalitetsoversikt.R
\name{mod_kvalitetsoversikt_ui}
\alias{mod_kvalitetsoversikt_ui}
\alias{mod_kvalitetsoversikt_server}
\title{mod_kvalitetsoversikt_ui and mod_kvalitetsoversikt_server}
\usage{
mod_kvalitetsoversikt_ui(id)

mod_kvalitetsoversikt_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
