% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar_qo.R
\name{mod_sidebar_qo_ui}
\alias{mod_sidebar_qo_ui}
\alias{mod_sidebar_qo_server}
\title{sidebar}
\usage{
mod_sidebar_qo_ui(id)

mod_sidebar_qo_server(input, output, session, register_data_description)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
