% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_navbar.R
\name{top_navbar_ui}
\alias{top_navbar_ui}
\alias{top_navbar_server}
\title{top navbar}
\usage{
top_navbar_ui(id)

top_navbar_server(input, output, session, app_data, config)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{app_data}{provide the treatment units}

\item{config}{configuration file}
}
\description{
A shiny Module.
}
\keyword{internal}
