% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national_table.R
\name{national_register_content}
\alias{national_register_content}
\title{constructs the body of a table}
\usage{
national_register_content(input_data, indicator_description, config)
}
\arguments{
\item{input_data}{a dataframe to construct the
table from}

\item{indicator_description}{the indicator description file}

\item{config}{the configuration setup}
}
\value{
tbody of the html table
}
\description{
constructs the body of a table
}
