% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national_table.R
\name{national_table}
\alias{national_table}
\title{national values in indicator table}
\usage{
national_table(input_data, indicator_description, config)
}
\arguments{
\item{input_data}{a dataframe to construct the
table from}

\item{indicator_description}{the indicator description file}

\item{config}{the configuration setup}
}
\description{
html table that displays the national values
in the indicator rows. The table is shown when a treatment
unit is not selected
}
