% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qi_overview.R
\name{overview_list}
\alias{overview_list}
\title{Makes html list of clickable links}
\usage{
overview_list(id, category_name, nr_of_reg)
}
\arguments{
\item{id}{unique id for each link}

\item{category_name}{the names of a category that will
be displayed}

\item{nr_of_reg}{the nr of registers included in a category}
}
\value{
html tags
}
\description{
Makes html list of clickable links
}
