% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_overview.R
\name{quality_overview}
\alias{quality_overview}
\alias{quality_overview_ui}
\alias{quality_overview_server}
\title{Quality overview}
\usage{
quality_overview_ui(id)

quality_overview_server(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
Quality overview module server, user interface and server functions.
Present results of search to user as table
}
\keyword{internal}
