% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national_table.R
\name{register_tr}
\alias{register_tr}
\title{adds table tags around register name}
\usage{
register_tr(register_name, col_span)
}
\arguments{
\item{register_name}{name of the register}

\item{col_span}{the total number of columns in the table}
}
\value{
a register name with tr and td tags including colSpan
}
\description{
A function that adds tr and td tags to the register
name with col_span.
}
