% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run}
\usage{
run_app()
}
\description{
Run the Shiny Application, use app_ui and app_server as
ui and server respectively
}
\seealso{
\code{\link{app_ui}}, \code{\link{app_server}}
}
