% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar_qo.R
\name{sidebar_qo_ui}
\alias{sidebar_qo_ui}
\alias{sidebar_qo_server}
\title{sidebar}
\usage{
sidebar_qo_ui(id)

sidebar_qo_server(id, register_data_description)
}
\arguments{
\item{id}{shiny id}

\item{register_data_description}{register_data_description}
}
\description{
A shiny Module.
}
\keyword{internal}
