% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_legend.R
\name{table_legend}
\alias{table_legend}
\alias{table_legend_ui}
\alias{table_legend_server}
\title{Table legend}
\usage{
table_legend_ui(id, config)

table_legend_server(id)
}
\arguments{
\item{id}{shiny id}

\item{config}{configuration file}
}
\description{
Table legend module server, user interface and server functions.
Buttons to filter by level of precision (maaloppnaaelse),
low, medium or high
}
\keyword{internal}
