% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{table_row_constructor}
\alias{table_row_constructor}
\title{constructs a row of a html table}
\usage{
table_row_constructor(dataframe_row)
}
\arguments{
\item{dataframe_row}{a row of a df}
}
\value{
a row of html table
}
\description{
constructs a row of a html table
}
