% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_navbar.R
\name{top_navbar}
\alias{top_navbar}
\alias{top_navbar_ui}
\alias{top_navbar_server}
\title{Top navbar}
\usage{
top_navbar_ui(id)

top_navbar_server(id, app_data, config)
}
\arguments{
\item{id}{shiny id}

\item{app_data}{provide the treatment units}

\item{config}{configuration file}
}
\description{
Table navigation bar
module server, user interface and server functions.
Lets user search for RHFs, HFs or hospitals, look at specific
years and get version information about the app
}
\keyword{internal}
