% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_info.R
\name{app_info}
\alias{app_info}
\alias{version_info}
\alias{no_opt_out_ok}
\title{Functions providing informational text about the app(s)}
\usage{
version_info(newline = "<br>")

no_opt_out_ok()
}
\arguments{
\item{newline}{String element defining line break for formatting. Default is
\code{<br>}}
}
\value{
String formatted in some sensible manner
}
\description{
Functions providing informational text about the app(s)
}
\examples{
version_info()
no_opt_out_ok()
}
