% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{compute_indicator_median}
\alias{compute_indicator_median}
\title{calculates median}
\usage{
compute_indicator_median(grouped_data)
}
\arguments{
\item{grouped_data}{grouped data}
}
\value{
aggrgated data with indicators
}
\description{
calculates median
}
