% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_source.R
\name{data_source}
\alias{data_source}
\alias{agg_data}
\title{Provide data for qmongr}
\usage{
agg_data()
}
\value{
List of data sets
}
\description{
Provide data sets used by qmongr. Currently based on context (env vars)
}
