% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_config}
\alias{get_config}
\title{Retrieve Config}
\usage{
get_config(config_file)
}
\arguments{
\item{config_file}{Config file to extract (yaml)}
}
\description{
Retrieves config file.
}
\keyword{internal}
