% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{load_data}
\alias{load_data}
\title{load data}
\usage{
load_data(data_type = "both")
}
\arguments{
\item{data_type}{the dataframe to be loaded.
It can be "description", "indicator" or "both".}
}
\value{
a list with dataframes
}
\description{
load data
}
\examples{
indicator_description <- load_data(data_type = "description")
indicator_data <- load_data(data_type = "indicator")
indicator_with_description <- load_data()

}
