% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{load_data}
\alias{load_data}
\title{Load data}
\usage{
load_data(data_type = "all")
}
\arguments{
\item{data_type}{the dataframes to be loaded.
It can be "description", "indicator", "hospital_name_structure"
and "all".}
}
\value{
a list with dataframes
}
\description{
Load data from qmongrdata
}
\examples{
indicator_description <- load_data(data_type = "description")
indicator_data <- load_data(data_type = "indicator")
indicator_with_description <- load_data()

}
