% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{qi_table}
\alias{qi_table}
\title{html table}
\usage{
qi_table(table_data, selected_units, config)
}
\arguments{
\item{table_data}{a dataframe to construct a html table from}

\item{config}{the configuration setup}

\item{unit}{the chosen unit}
}
\description{
html table
}
