% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_overview.R
\name{quality_overview_ui}
\alias{quality_overview_ui}
\alias{quality_overview_server}
\title{quality_overview_ui and quality_overview_server}
\usage{
quality_overview_ui(id)

quality_overview_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
