% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar_qo.R
\name{sidebar_qo}
\alias{sidebar_qo}
\alias{sidebar_qo_ui}
\alias{sidebar_qo_server}
\title{Sidebar quality overview}
\usage{
sidebar_qo_ui(id, list_of_med_fields)

sidebar_qo_server(id, register_data_description)
}
\arguments{
\item{id}{shiny id}

\item{list_of_med_fields}{List of medical fields, with number of indicators per field.}

\item{register_data_description}{register_data_description}
}
\description{
Sidebar quality overview module server, user interface and server functions.
Lets user filter by medical field
}
\keyword{internal}
