% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{table_body_constructor}
\alias{table_body_constructor}
\title{Construct table body}
\usage{
table_body_constructor(
  datatable,
  units,
  national,
  indicator_description,
  config
)
}
\arguments{
\item{datatable}{a row of a df}

\item{units}{selected treatment units}

\item{national}{a df with national indicator values}

\item{config}{the configuration setup}
}
\value{
a row of html table
}
\description{
Constructs the body of a table, used by qi_table
}
\seealso{
\code{\link{qi_table}}
}
