% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{table_data}
\alias{table_data}
\title{insert values to a td tag}
\usage{
table_data(units, table_cell_data, indicator_name)
}
\arguments{
\item{units}{selected treatment units}

\item{table_cell_data}{table cell data}

\item{indicator_name}{names of the input qi}
}
\value{
the contnets of a td tag
}
\description{
insert values to a td tag
}
