% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_legend.R
\name{table_legend_ui}
\alias{table_legend_ui}
\alias{table_legend_server}
\title{Table legend}
\usage{
table_legend_ui(id, config)

table_legend_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{config}{configuration file}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
